Reorganized config files (eslint, jestrc, postcss.config [-> also refactored]) and updated package.json | webpacks
moved importing chai methods into setupTest.js
